Changing CMakeLists for noraml work on windows Adding all files to CMakeLists automatickly Fixing idle timer Some small code optimisations Renaming some files

+++ src/data/initFile.hpp
@@ -13,9 +13,9 @@
 class InitFile {
  public:
     Uint8 language;           // Game language for all texts
-    Uint8 musicVolume;        // Start volume of music
-    Uint8 soundsVolume;       // Start volume of effects
-    Uint16 drawFPS;           // Setted frames per seconds
+    unsigned musicVolume;     // Start volume of music
+    unsigned soundsVolume;    // Start volume of effects
+    unsigned drawFPS;         // Setted frames per seconds
     std::string baseIP;       // Saved ip for better expirience
     std::string basePort;     // Saved connection port for better expirience
     std::string startConfig;  // Start field configuration
